[git-webkit] Allow independent installation of hooks https://bugs.webkit.org/show_bug.cgi?id=253897 rdar://106716871
Reviewed by Elliott Williams.

Move hook installation into an independent program so services can install specific hooks.

* Tools/Scripts/libraries/webkitscmpy/setup.py: Bump version.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/__init__.py: Ditto.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/__init__.py: Include InstallHooks program.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/install_hooks.py: Added.
(InstallHooks.parser): Allow caller to specify default mode and specific hook to install.
(InstallHooks._security_levels): Moved from Setup._security_levels.
(InstallHooks.main): Moveed hook installation from Setup.git.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/program/setup.py:
(Setup.git): Move hook installation to InstallHooks.
(Setup._security_levels): Move to InstallHooks._security_levels.
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/test/setup_unittest.py:

Canonical link: https://commits.webkit.org/261653@main
