+++ Tools/Scripts/libraries/webkitscmpy/setup.py
@@ -29,7 +29,7 @@ def readme():
 
 setup(
     name='webkitscmpy',
-    version='6.0.0',
+    version='6.0.1',
     description='Library designed to interact with git and svn repositories.',
     long_description=readme(),
     classifiers=[
